+++ packages/desktop/.gitignore
