+++ packages/desktop/package.json
