+++ packages/desktop/public/kalzate-icon.ico
