+++ packages/desktop/public/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "1.0.0",
   "description": "Serverless Point Of Sale App For Humans",
-  "main": "index.prod.js",
+  "main": "index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/kalzate/kalzate.git"
