+++ packages/desktop/public/print_receipt.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+  <html>
+  <head>
+    <meta charset="utf-8">
+    <style>
+      @page {
+        margin: 0;
+      }
+      body * {
+        visibility: hidden;
+        margin: 0;
+      }
+      #printer-content, #printer-content * {
+        visibility: visible;
+      }
+      #printer-content {
+        position: absolute;
+        left: 0;
+        top: 0;
+        width: 71mm
+        min-width: 71mm;
+        max-width: 71mm;
+        font-size: 11px;
+        font-family: monospace, "Courier New", Courier !IMPORTANT;
+        font-weight: 400;
+      }
+    </style>
+  </head>
+  <body><pre id="printer-content">
+MY STORE Kalzate
+
+Address: YYY
+Email: test@test.com
+Phone: +34XXX-XXX-XXX
+Date: 10/07/2020 13:04
+
+Ticket: 1594379063206
+Category: Ticket
+
+--------------------------------------------
+Item                   U.   PSP    Subtotal
+--------------------------------------------
+1600x 3.6G             1    135.95 135.95 
+--------------------------------------------
+                            TOTAL: 135.95€
+
+
+Amount: 0.00€,
+Return: -135.95€
+
+RETURNS ADMITTED BETWEEN 10/07/2020 AND 30/07/2020 
+
+*** THANKS FOR SHOPPING AT Kalzate ***
+
+-   -   -   -   -   -   -   -   -   -   -
+</pre></body>
+  </html>
