+++ packages/desktop/src/index-splash-loading.html
