+++ packages/desktop/src/index.js
@@ -7,13 +7,12 @@ const ipc = electron.ipcMain;
 const BrowserWindow = electron.BrowserWindow;
 const path = require('path');
 const url = require('url');
-const fs = require('fs');
-const os = require('os');
-// const Printer = require('electron-printer');
+const printTicket = require('./printer');
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let mainWindow;
+// let splashWindow;
 
 function createWindow() {
   // Create the browser window.
@@ -31,9 +30,18 @@ function createWindow() {
   });
 
   mainWindow.once('ready-to-show', () => {
+    // splashWindow.destroy();
     mainWindow.maximize();
     mainWindow.show();
   });
+
+  // splashWindow = new BrowserWindow({ width: 300, height: 300, transparent: true, frame: false, alwaysOnTop: true });
+  // splashWindow.loadURL(url.format({
+  //   pathname: path.join(__dirname, 'index-splash-loading.html'),
+  //   protocol: 'file:',
+  //   slashes: true,
+  // }));
+
 
   // and load the index.html of the app.
   mainWindow.loadURL(
@@ -79,86 +87,4 @@ app.on('activate', () => {
 });
 
 
-// function print(text, options) {
-//   const toBytes = (str) => Array.from(str).map((c) => c.charCodeAt(0));
-
-//   const printData = toBytes(text).concat([0x1B, 0x69, 0x1B, 0x70, 0x00, 0x09, 0x09]);
-
-//   require('kalzate-printer').print({
-//     text: new Buffer(printData),
-//     printer: 'termica',
-//   });
-// }
-
-/**
- * This method relies on a native approach, but it limited as it will
- * not work using codes (to open cash drawer or cut the paper) and also
- * it has to be tweak for each printer. It also requires to set the
- * receipt printer as the default printer
- * @param {*} text the ticket to print
- */
-function printNative(text) {
-  const printFile = path.resolve(path.normalize(path.join(os.tmpdir(), 'print.html')));
-
-  fs.writeFileSync(printFile, `
-  <!DOCTYPE html>
-  <html>
-  <head>
-    <meta charset="utf-8">
-    <style>
-    @media print {
-      @page {
-        margin: 0;
-      }
-      body * {
-        visibility: hidden;
-      }
-      #printer-content, #printer-content * {
-        visibility: visible;
-      }
-      #printer-content {
-        position: absolute;
-        left: 0;
-        top: 0;
-        width: 71mm
-        min-width: 71mm;
-        max-width: 71mm;
-        font-size: 10px;
-        font-family: monospace, "Courier New", Courier !IMPORTANT;
-        font-weight: 600;
-      }
-    }
-    </style>
-  </head>
-  <body><pre id="printer-content">${text}</pre></body>
-  </html>
-  `);
-
-  const win = new BrowserWindow({ show: false });
-
-  win.loadURL(
-    url.format({
-      pathname: printFile,
-      protocol: 'file:',
-      slashes: true,
-    })
-  );
-
-  win.webContents.on('did-finish-load', () => {
-    win.webContents.print({ silent: true });
-    setTimeout(() => {
-      win.close();
-    }, 1000);
-  });
-}
-
-
-ipc.on('print-ticket', (event, text, options) => {
-  try {
-    // print(text, options);
-    printNative(text, options);
-  } catch (error) {
-    console.error(error);
-    // printNative(text, options);
-  }
-});
+ipc.on('print-ticket', (_, text, options) => printTicket(text, options));
