--- packages/electron/src/index.prod.js
@@ -1,108 +1,0 @@
-const electron = require('electron');
-// Module to control application life.
-const app = electron.app;
-// IPC communication
-const ipc = electron.ipcMain;
-// Module to create native browser window.
-const BrowserWindow = electron.BrowserWindow;
-const path = require('path');
-const url = require('url');
-const fs = require('fs');
-const os = require('os');
-
-// Keep a global reference of the window object, if you don't, the window will
-// be closed automatically when the JavaScript object is garbage collected.
-let mainWindow;
-let splashWindow;
-
-function createWindow() {
-  // Create the browser window.
-  mainWindow = new BrowserWindow({
-    show: false,
-    frame: false,
-    height: 600,
-    width: 800,
-    'min-height': 500,
-    'min-width': 900,
-    webPreferences: {
-      nodeIntegration: false,
-      preload: path.join(__dirname, 'preload.js'),
-    },
-  });
-
-  mainWindow.once('ready-to-show', () => {
-    splashWindow.destroy();
-    mainWindow.maximize();
-    mainWindow.show();
-  });
-
-  splashWindow = new BrowserWindow({ width: 300, height: 300, transparent: true, frame: false, alwaysOnTop: true });
-  splashWindow.loadURL(url.format({
-    pathname: path.join(__dirname, 'index-splash-loading.html'),
-    protocol: 'file:',
-    slashes: true,
-  }));
-
-
-  // and load the index.html of the app.
-  mainWindow.loadURL(
-    url.format({
-      pathname: path.join(__dirname, 'index.html'),
-      protocol: 'file:',
-      slashes: true,
-    })
-  );
-
-  // Open the DevTools.
-  // mainWindow.webContents.openDevTools();
-
-  // Emitted when the window is closed.
-  mainWindow.on('closed', () => {
-    // Dereference the window object, usually you would store windows
-    // in an array if your app supports multi windows, this is the time
-    // when you should delete the corresponding element.
-    mainWindow = null;
-  });
-}
-
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
-
-// Quit when all windows are closed.
-app.on('window-all-closed', () => {
-  // On OS X it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-  if (process.platform !== 'darwin') {
-    app.quit();
-  }
-});
-
-app.on('activate', () => {
-  // On OS X it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (mainWindow === null) {
-    createWindow();
-  }
-});
-
-ipc.on('print-ticket', (event, text) => {
-  try {
-    const printFile = path.resolve(path.normalize(path.join(os.tmpdir(), 'print.txt')));
-    fs.writeFileSync(printFile, text);
-    const win = new BrowserWindow({ show: false });
-    win.loadURL(`file://${printFile}`);
-    win.webContents.on('did-finish-load', () => {
-      win.webContents.print({ silent: true });
-      setTimeout(() => {
-        win.close();
-      }, 1000);
-    });
-  } catch (error) {
-    console.error(error);
-  }
-});
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
